+++ src/wallet/infrastructure/repository/protocols.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional, Protocol, Self
 import uuid
 
@@ -6,7 +7,7 @@ from wallet.domain.wallet import Wallet
 
 class WalletRepositoryProtocol(Protocol):
     async def update_wallet_balance(
-        self: Self, wallet_id: uuid.UUID, new_balance: float
+        self: Self, wallet_id: uuid.UUID, new_balance: Decimal
     ): ...
 
     async def get_wallet_for_update(
